Split URNG stuff out of alg/shuffle.hpp into new util/random.hpp
