+++ range-v3.vcxproj.filters
@@ -1088,6 +1088,9 @@
     <ClInclude Include="include\range\v3\utility\optional.hpp">
       <Filter>Header Files\range\v3\utility</Filter>
     </ClInclude>
+    <ClInclude Include="include\range\v3\utility\random.hpp">
+      <Filter>Header Files\range\v3\utility</Filter>
+    </ClInclude>
     <ClInclude Include="include\range\v3\utility\safe_int.hpp">
       <Filter>Header Files\range\v3\utility</Filter>
     </ClInclude>
